Add support for sharp edges
Now, you can have edges specified as sharp and they will not be merged,
allowing them to be shown as sharp inside the game. Previously vertices
that were in the same location were merged regardless.

To use, mark edges as sharp and then split those edges. This creates
enough vertices for the game to have different vertex normals, rendering
the faces as flat.